refactor: use sampleTerrainHeightFromCartesian
